refactor(rewrite): split reorder_static_placement into subpackage
Convert the flat module into a subpackage with one file per concern:

  types.py  — _BARRIERS, _SchedulableStmt, _group_key, _group_within_layer
  asap.py   — _asap_schedule, asap_reorder_policy
  alap.py   — _alap_schedule, alap_reorder_policy
  rule.py   — ReorderStaticPlacement rewrite rule
  __init__.py — re-exports the three public names

All existing imports (passes.py, tests) are unchanged.
