--- python/bloqade/lanes/rewrite/reorder_static_placement.py
@@ -1,235 +1,0 @@
-from dataclasses import dataclass
-from typing import Callable
-
-import rustworkx
-from kirin import ir
-from kirin.rewrite import abc
-
-from bloqade.lanes.dialects import place
-from bloqade.lanes.types import StateType
-
-_BARRIERS: tuple[type, ...] = (place.Initialize, place.EndMeasure)
-
-# Union of all concrete QuantumStmt subclasses that carry a .qubits attribute.
-_SchedulableStmt = (
-    place.R | place.Rz | place.StarRz | place.CZ | place.Initialize | place.EndMeasure
-)
-
-
-def _group_key(stmt: _SchedulableStmt) -> tuple:
-    """Hashable key identifying gates that FuseAdjacentGates can fuse together."""
-    if isinstance(stmt, place.R):
-        return (type(stmt), id(stmt.axis_angle), id(stmt.rotation_angle))
-    if isinstance(stmt, place.Rz):
-        return (type(stmt), id(stmt.rotation_angle))
-    if isinstance(stmt, place.StarRz):
-        return (type(stmt), id(stmt.rotation_angle), stmt.qubit_indices)
-    return (type(stmt),)  # CZ has no non-qubit params
-
-
-def _group_within_layer(layer_stmts: list[_SchedulableStmt]) -> list[_SchedulableStmt]:
-    """Re-order layer_stmts so fusable-equivalent gates are adjacent.
-
-    Uses Python dict insertion-order: the first time a (type, params) key is
-    seen a new group is opened; subsequent matches append to that group.
-    Iterating over groups in insertion order preserves first-seen ordering.
-    """
-    groups: dict[tuple, list[_SchedulableStmt]] = {}
-    for stmt in layer_stmts:
-        groups.setdefault(_group_key(stmt), []).append(stmt)
-    return [stmt for group in groups.values() for stmt in group]
-
-
-def _asap_schedule(stmts: list[_SchedulableStmt]) -> list[_SchedulableStmt]:
-    """Return stmts in ASAP layer order with fusable-equivalent gates adjacent within each layer."""
-    if len(stmts) <= 1:
-        return list(stmts)
-
-    dag = rustworkx.PyDAG()
-    node_for: list[int] = [dag.add_node(i) for i in range(len(stmts))]
-    last_touch: dict[int, int] = {}
-
-    for i, stmt in enumerate(stmts):
-        for q in stmt.qubits:
-            if q in last_touch:
-                dag.add_edge(last_touch[q], node_for[i], None)
-            last_touch[q] = node_for[i]
-
-    layer: dict[int, int] = {}
-    for node_idx in rustworkx.topological_sort(dag):
-        preds = dag.predecessor_indices(node_idx)
-        layer[node_idx] = 0 if not preds else max(layer[p] for p in preds) + 1
-
-    # Bucket statements per layer in original order, then group within each layer
-    # so fusable-equivalent gates (same type + params) are adjacent.
-    layer_buckets: dict[int, list[_SchedulableStmt]] = {}
-    for i, stmt in enumerate(stmts):
-        layer_buckets.setdefault(layer[node_for[i]], []).append(stmt)
-
-    result: list[_SchedulableStmt] = []
-    for lyr in sorted(layer_buckets):
-        result.extend(_group_within_layer(layer_buckets[lyr]))
-    return result
-
-
-def asap_reorder_policy(
-    stmts: list[_SchedulableStmt],
-) -> list[_SchedulableStmt]:
-    """ASAP scheduling policy. Hard barriers (Initialize, EndMeasure) divide
-    the body into independent segments; each segment is scheduled separately."""
-    result: list[_SchedulableStmt] = []
-    segment: list[_SchedulableStmt] = []
-
-    for stmt in stmts:
-        if isinstance(stmt, _BARRIERS):
-            result.extend(_asap_schedule(segment))
-            result.append(stmt)
-            segment = []
-        else:
-            segment.append(stmt)
-    result.extend(_asap_schedule(segment))
-    return result
-
-
-def _alap_schedule(stmts: list[_SchedulableStmt]) -> list[_SchedulableStmt]:
-    """Return stmts in ALAP layer order with fusable-equivalent gates adjacent within each layer.
-
-    Each gate is assigned to the *latest* layer that still respects all data
-    dependencies (i.e. every successor is placed at a strictly later layer).
-    Deferring single-qubit gates reduces the qubit footprint of early
-    CZ-anchored StaticPlacement regions, lowering atom-move overhead.
-    """
-    if len(stmts) <= 1:
-        return list(stmts)
-
-    dag = rustworkx.PyDAG()
-    node_for: list[int] = [dag.add_node(i) for i in range(len(stmts))]
-    last_touch: dict[int, int] = {}
-
-    for i, stmt in enumerate(stmts):
-        for q in stmt.qubits:
-            if q in last_touch:
-                dag.add_edge(last_touch[q], node_for[i], None)
-            last_touch[q] = node_for[i]
-
-    topo_order = list(rustworkx.topological_sort(dag))
-
-    # Forward pass to find max ASAP depth (= ALAP horizon for sink nodes).
-    asap_layer: dict[int, int] = {}
-    for node_idx in topo_order:
-        preds = dag.predecessor_indices(node_idx)
-        asap_layer[node_idx] = 0 if not preds else max(asap_layer[p] for p in preds) + 1
-
-    max_layer = max(asap_layer.values()) if asap_layer else 0
-
-    # Backward pass: ALAP[i] = min(ALAP[j] for j in successors(i)) - 1,
-    # or max_layer if i has no successors.
-    alap_layer: dict[int, int] = {}
-    for node_idx in reversed(topo_order):
-        succs = dag.successor_indices(node_idx)
-        alap_layer[node_idx] = (
-            max_layer if not succs else min(alap_layer[s] for s in succs) - 1
-        )
-
-    layer_buckets: dict[int, list[_SchedulableStmt]] = {}
-    for i, stmt in enumerate(stmts):
-        layer_buckets.setdefault(alap_layer[node_for[i]], []).append(stmt)
-
-    result: list[_SchedulableStmt] = []
-    for lyr in sorted(layer_buckets):
-        result.extend(_group_within_layer(layer_buckets[lyr]))
-    return result
-
-
-def alap_reorder_policy(
-    stmts: list[_SchedulableStmt],
-) -> list[_SchedulableStmt]:
-    """ALAP scheduling policy. Hard barriers (Initialize, EndMeasure) divide
-    the body into independent segments; each segment is scheduled separately.
-
-    Gates are assigned to their *latest* valid layer, deferring single-qubit
-    gates as long as possible to minimise the qubit footprint of earlier
-    CZ-anchored StaticPlacement regions.
-    """
-    result: list[_SchedulableStmt] = []
-    segment: list[_SchedulableStmt] = []
-
-    for stmt in stmts:
-        if isinstance(stmt, _BARRIERS):
-            result.extend(_alap_schedule(segment))
-            result.append(stmt)
-            segment = []
-        else:
-            segment.append(stmt)
-    result.extend(_alap_schedule(segment))
-    return result
-
-
-@dataclass
-class ReorderStaticPlacement(abc.RewriteRule):
-    """Reorder quantum statements within a StaticPlacement using a pluggable policy.
-
-    The policy receives all schedulable statements from the body (R, Rz,
-    StarRz, CZ, Initialize, EndMeasure — everything except the trailing Yield)
-    and returns them in the desired order.  Barrier handling (Initialize,
-    EndMeasure) is the policy's responsibility; ``asap_reorder_policy`` segments
-    on barriers and schedules each segment independently.
-
-    If the body contains any statement type outside that supported set the
-    rewriter skips the node rather than silently dropping unknown statements.
-    """
-
-    reorder_policy: Callable[[list[_SchedulableStmt]], list[_SchedulableStmt]]
-
-    def rewrite_Statement(self, node: ir.Statement) -> abc.RewriteResult:
-        if not isinstance(node, place.StaticPlacement):
-            return abc.RewriteResult()
-
-        body_block = node.body.blocks[0]
-        old_yield = body_block.last_stmt
-        assert isinstance(old_yield, place.Yield)
-
-        _supported = (
-            place.R,
-            place.Rz,
-            place.StarRz,
-            place.CZ,
-            place.Initialize,
-            place.EndMeasure,
-        )
-        stmts: list[_SchedulableStmt] = []
-        for s in body_block.stmts:
-            if isinstance(s, place.Yield):
-                continue
-            if not isinstance(s, _supported):
-                return abc.RewriteResult()
-            stmts.append(s)
-
-        if not stmts:
-            return abc.RewriteResult()
-
-        new_stmts = self.reorder_policy(stmts)
-
-        if [id(s) for s in new_stmts] == [id(s) for s in stmts]:
-            return abc.RewriteResult()
-
-        new_body = ir.Region(new_block := ir.Block())
-        curr_state = new_block.args.append_from(StateType, "entry_state")
-
-        for stmt in new_stmts:
-            remapped = stmt.from_stmt(stmt, args=(curr_state, *stmt.args[1:]))
-            new_block.stmts.append(remapped)
-            curr_state = remapped.state_after
-            for old_r, new_r in zip(stmt.results[1:], remapped.results[1:]):
-                old_r.replace_by(new_r)
-
-        new_block.stmts.append(place.Yield(curr_state, *old_yield.classical_results))
-
-        new_sp = place.StaticPlacement(node.qubits, new_body)
-        new_sp.insert_before(node)
-
-        for old_r, new_r in zip(node.results, new_sp.results, strict=True):
-            old_r.replace_by(new_r)
-
-        node.delete()
-        return abc.RewriteResult(has_done_something=True)
