+++ python/bloqade/lanes/rewrite/reorder_static_placement/__init__.py
@@ -1,0 +1,9 @@
+from bloqade.lanes.rewrite.reorder_static_placement.alap import (
+    alap_reorder_policy as alap_reorder_policy,
+)
+from bloqade.lanes.rewrite.reorder_static_placement.asap import (
+    asap_reorder_policy as asap_reorder_policy,
+)
+from bloqade.lanes.rewrite.reorder_static_placement.rule import (
+    ReorderStaticPlacement as ReorderStaticPlacement,
+)
